separated index.css on a smaller files; removed unnecessary files
