+++ src/components/Header/Header.css
@@ -1,0 +1,67 @@
+.main__header {
+  background: linear-gradient(
+    to bottom right,
+    rgb(231 255 202),
+    rgb(239 126 184)
+  );
+  box-shadow: 0 12px 34px -11px rgb(38 20 19 / 10%);
+  border-bottom: 2px solid rgba(0, 0, 0, 0.3);
+  backdrop-filter: blur(2px);
+  font-weight: 500;
+  font-size: 1.3rem;
+  z-index: 99;
+}
+
+.header__nav {
+  color: var(--main-color)!important;
+}
+
+.header__nav__links a.nav-link,
+.header__nav__links button.nav-link,
+.header__nav__user-btns a.nav-link,
+.header__nav__user-btns button.nav-link {
+  padding: .5rem;
+  margin: 0;
+}
+
+.header__nav__user-btns a.nav-link {
+  opacity: .8;
+}
+
+.header__nav__user-btns a.nav-link:hover,
+.header__nav__user-btns a.nav-link.active {
+  opacity: 1;
+}
+
+.header__nav__links a.nav-link.active {
+  margin-bottom: -0.2rem;
+  border-bottom: 0.2rem solid var(--main-color);
+}
+
+.cart__couter-icon {
+  background-color: var(--btn-main-color);
+}
+
+.navbar__brand-text {
+  font-family: 'Comfortaa', cursive;
+  font-size: 1.4rem;
+}
+
+.header__user-avatar {
+  border: 3px solid var(--main-color);
+  box-shadow: 0 0px 6px .5px rgba(28, 34, 46, .3);
+  width: 3rem;
+  height: 3rem;
+}
+
+.navbar-toggler{
+ border: none;
+}
+
+.navbar-toggler:focus {
+  box-shadow: none;
+}
+
+.dropdown-toggle, .dropdown-toggle:active {
+  border: none;
+}
