+++ src/components/Pages/Login/Login.css
@@ -1,0 +1,15 @@
+.login__card {
+  margin: 0;
+  background: rgba(235, 248, 221, .4);
+  background: linear-gradient(
+    to bottom right,
+    rgba(235, 248, 221, .4),
+    rgba(255, 200, 229, .4)
+  );
+}
+
+.login__alert {
+  padding: 6px 12px;
+  margin: 6px 0;
+}
+
