+++ src/components/Pages/Login/Login.jsx
@@ -1,4 +1,5 @@
 import { Outlet } from 'react-router-dom';
+import './Login.css';
 
 export function Login() {
   return (
