+++ src/index.js
@@ -11,7 +11,6 @@ import {
 } from '@tanstack/react-query';
 import { Provider } from 'react-redux';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import { ErrorPage } from './components/Pages/Error/Error';
 import { Products } from './components/Pages/Products/Products';
 import { Main } from './components/Pages/Main/Main';
@@ -96,8 +95,3 @@ root.render(
     </QueryClientProvider>
   </React.StrictMode>,
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
